feat, fix: added dotenv, content adjustment, navigation fix

+++ backend/controllers/adminControllers/editNews.controller.js
@@ -3,17 +3,22 @@ const fs = require('fs')
 const { updateNews } = require('../../services/news/news.service')
 
 async function editNewsController(req, res) {
-    const { id, header, text, tags } = req.body;
+    const { id, shortHeader, header, text, tags } = req.body;
     const photo = req.file;
-    console.log(id, header, text, tags)
+    console.log(photo)
+    // console.log(id, shortHeader,header, text, tags)
     try {
-        const result = await updateNews(id, header, text, tags, photo.filename)
-        console.log(result.oldFilename.dataValues.name)
-        const oldFilename = result.oldFilename.dataValues.name;
-        fs.unlink(`./static/${oldFilename}`, (err) => {
-            console.log(err)
-        });
-        res.status(200).json(result)
+        if (photo){
+            const result = await updateNews(id, shortHeader, header, text, tags, photo.filename);
+            const oldFilename = result.oldFilename.dataValues.filename;
+            fs.unlink(`./static/${oldFilename}`, (err) => {
+                console.log(err)
+            });
+            res.status(200).json(result)
+        } else {
+            const result = await updateNews(id, shortHeader, header, text, tags);
+            res.status(200).json(result)
+        }
     } catch (error) {
         console.log('Catched error')
         console.log(error)
